requirements.txt
numpy
pandas
matplotlib
scikit-learn
finnhub-python
